Created separate file for the Account class
